Adds JtiMatcher revocation strategy
This strategy must be included in the user model, and requires that it
has a `jti` column. It adds the value of the `jti` column as the `jti`
claim in dispatched tokens.

In order to tell whether a token is revoked, it just compares both `jti`
values. On revocation, it changes column value so that the token is no
longer valid.